pyproject.toml
[tool.poetry]
name = "dadaia-tools"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
redis = "^5.0.0"
kafka-python = "^2.0.2"
azure-data-tables = "^12.4.4"
azure-keyvault-secrets = "^4.7.0"
azure-storage-blob = "^12.18.2"
boto3 = "^1.28.57"
azure-identity = "^1.14.0"
azure-storage-queue = "^12.7.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
blue = "^0.9.1"
isort = "^5.12.0"
taskipy = "^1.12.0"
python-dotenv = "^1.0.0"
faker = "^19.6.2"


[tool.poetry.group.doc.dependencies]
mkdocstrings = "^0.23.0"
mkdocstrings-python = "^1.7.0"
mkdocs-material = "^9.3.1"
pymdown-extensions = "^10.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"
markers = [
    "a_verificar: a_verificar",
    "redis_client: redis_client",
    "kafka_admin: kafka_admin",
    "kafka_client: kafka_client",
    "blob_admin: blob_admin",
    "blob_client: blob_client",
    "key_vault_client: key_vault_client",
    "table_client: table_client",
    "queue_client: queue_client",
    "etherscan_client: etherscan_client",
    "latest: latest",
]

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
docs = "mkdocs serve"
test = "pytest --cov=dadaia_tools -vv"
post_test = "coverage html && firefox htmlcov/index.html"
